back to running all tests .. on dongle

+++ src/platforms/nRF5/channel-serial.c
@@ -16,7 +16,7 @@ static int  data_available=0;
 
 static channel_serial_connect_cb connect_cb;
 
-void channel_serial_on_recv(char* ch, int len)
+void channel_serial_on_recv(char* ch, uint16_t len)
 {
   if(!ch){
     if(connect_cb) connect_cb("serial");
